making everything work for python 2.7
