+++ sdk/tables/azure-data-tables/tests/test_table_batch_async.py
@@ -38,7 +38,7 @@ from azure.data.tables import (
 )
 
 from _shared.asynctestcase import AsyncTableTestCase
-from preparers import tables_decorator_async
+from async_preparers import tables_decorator_async
 
 #------------------------------------------------------------------------------
 TEST_TABLE_PREFIX = 'table'
