+++ sdk/tables/azure-data-tables/tests/test_table_cosmos_async.py
@@ -21,7 +21,7 @@ from azure.data.tables.aio import TableServiceClient
 
 from _shared.asynctestcase import AsyncTableTestCase
 from _shared.testcase import SLEEP_DELAY
-from preparers import cosmos_decorator_async
+from async_preparers import cosmos_decorator_async
 
 TEST_TABLE_PREFIX = 'pytableasync'
