+++ sdk/tables/azure-data-tables/tests/test_table_entity_async.py
@@ -35,7 +35,7 @@ from azure.data.tables import (
 from azure.data.tables.aio import TableServiceClient
 
 from _shared.asynctestcase import AsyncTableTestCase
-from preparers import tables_decorator_async
+from async_preparers import tables_decorator_async
 
 class StorageTableEntityTest(AzureTestCase, AsyncTableTestCase):
